+++ app/src/main/java/com/scipath/becomeaking/view/layout/CityMarkerLayout.java
@@ -10,28 +10,28 @@ import androidx.annotation.Nullable;
 import com.scipath.becomeaking.R;
 
 
-public class CityLayout extends LinearLayout {
+public class CityMarkerLayout extends LinearLayout {
 
     TextView textViewCityName;
 
 
-    public CityLayout(Context context) {
+    public CityMarkerLayout(Context context) {
         super(context);
         init(context);
     }
 
-    public CityLayout(Context context, @Nullable AttributeSet attrs) {
+    public CityMarkerLayout(Context context, @Nullable AttributeSet attrs) {
         super(context, attrs);
         init(context);
     }
 
-    public CityLayout(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+    public CityMarkerLayout(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
         init(context);
     }
 
     private void init(Context context) {
-        inflate(context, R.layout.layout_city, this);
+        inflate(context, R.layout.layout_city_marker, this);
         textViewCityName = findViewById(R.id.text_view_city_name);
     }