+++ app/src/main/java/com/scipath/becomeaking/view/layout/PersonageMarkerLayout.java
@@ -1,0 +1,32 @@
+package com.scipath.becomeaking.view.layout;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.widget.LinearLayout;
+
+import androidx.annotation.Nullable;
+
+import com.scipath.becomeaking.R;
+
+
+public class PersonageMarkerLayout extends LinearLayout {
+
+    public PersonageMarkerLayout(Context context) {
+        super(context);
+        init(context);
+    }
+
+    public PersonageMarkerLayout(Context context, @Nullable AttributeSet attrs) {
+        super(context, attrs);
+        init(context);
+    }
+
+    public PersonageMarkerLayout(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+        init(context);
+    }
+
+    private void init(Context context) {
+        inflate(context, R.layout.layout_personage_marker, this);
+    }
+}